refactor: infrastructure and domains
