+++ src/domains/news/news-schema.test.ts
