+++ src/domains/news/news-schema.ts
